Work Manager Fixed

+++ app/src/main/java/com/screentimex/nouzze/Adapters/AppInfoListAdapter.kt
@@ -47,7 +47,7 @@ class AppInfoListAdapter(private val context: Context, private val appInfoList: 
         listViewHolder.appName.text = appInfo.appName
 //        listViewHolder.appIcon.setImageDrawable(appInfo.icon)
         listViewHolder.isTracked.visibility = View.VISIBLE
-        val time = showTimeSpent(appInfo.timeUseApp)
+        val time = showTimeSpent(appInfo.timeUseApp.toInt())
         listViewHolder.isTracked.text = time
         return convertView!!
     }
